bank_of_anthos_agent/requirements.txt
google-adk
gunicorn
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform
google-generativeai